Rename Service\System to Foundation\Environment\FunctionInspector
